@@ -5,40 +5,31 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import net.dv8tion.jda.api.entities.Member;
-import net.dv8tion.jda.api.entities.Role;
+import main.CommandManager;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 
 public abstract class Command {
 
-    protected String prefix = ".";
     private List<String> commandPrefixes = new ArrayList<>();
     private List<String> permission = new ArrayList<>();
     private List<Command> subCommands = new ArrayList<>();
     private String description = "";
     private String helpString = "";
     private String topic = "uncategorized";
+    private int minArguments = 0;
 
     public void handle(MessageReceivedEvent event, String[] argStrings) {
-        String[] cutArguments = cutArguments(argStrings, 1, argStrings.length);
-        if (Boolean.TRUE == isAllowed(event.getMember())) {
-            if (cutArguments.length < 1) {
-                handleImpl(event, cutArguments);
-            } else {
-                Command foundSubCommand = findSubCommand(cutArguments[0]);
-                if (foundSubCommand == null) {
-                    handleImpl(event, cutArguments);
-                } else {
-                    foundSubCommand.handle(event, cutArguments);
-                }
-            }
+        if (argStrings.length < minArguments) {
+            CommandManager.sendMessage(event, getHelp());
+        } else {
+            handleImpl(event, argStrings);
         }
     }
 
     public Command findSubCommand(String argument) {
         Command foundSubCommand = null;
         for (Command subCommand : this.subCommands) {
-            if (subCommand.getCommand().contains(this.prefix + argument)) {
+            if (subCommand.getCommand().contains(argument)) {
                 foundSubCommand = subCommand;
             }
         }
@@ -47,7 +38,9 @@ public abstract class Command {
 
     public String getHelp() {
         final StringBuilder finalHelpString = new StringBuilder();
-        finalHelpString.append(this.helpString).append("\n\n");
+        finalHelpString.append("***").append(getCommandHelpString()).append("***");
+        finalHelpString.append(" - ").append(getDescription()).append("\n");
+        finalHelpString.append(helpString.isBlank() ? "*No help available*" : helpString).append("\n\n");
         if (!this.subCommands.isEmpty()) {
             finalHelpString.append("available subcommands:").append("\n");
             this.subCommands.forEach(subCommand -> finalHelpString.append(subCommand.getCommandHelpString())
@@ -63,9 +56,7 @@ public abstract class Command {
     }
 
     protected void setHelpString(String helpString) {
-        this.helpString += "***" + getCommandHelpString() + "***";
-        this.helpString += " - " + getDescription() + "\n";
-        this.helpString += helpString;
+        this.helpString = helpString;
     }
 
     public List<Command> getSubCommands() {
@@ -82,7 +73,7 @@ public abstract class Command {
 
     protected void addCommendPrefix(String... commandPrefix) {
         List<String> newCommandPrefixes = Arrays.asList(commandPrefix).stream()
-                .map(newCommandPrefix -> prefix + newCommandPrefix).collect(Collectors.toList());
+                .map(newCommandPrefix -> newCommandPrefix).collect(Collectors.toList());
         this.commandPrefixes.addAll(newCommandPrefixes);
     }
 
@@ -91,8 +82,7 @@ public abstract class Command {
     }
 
     public void setCommand(String... cmd) {
-        commandPrefixes = Arrays.asList(cmd).stream().map(commandPrefix -> this.prefix + commandPrefix)
-                .collect(Collectors.toList());
+        commandPrefixes = Arrays.asList(cmd).stream().map(commandPrefix -> commandPrefix).collect(Collectors.toList());
     };
 
     public List<String> getCommand() {
@@ -115,27 +105,17 @@ public abstract class Command {
         this.topic = topic;
     }
 
-    public Boolean isAllowed(Member member) {
-        if (this.permission.contains("everyone"))
-            return true;
-        for (Role role : member.getRoles()) {
-            if (this.permission.contains(role.getName()))
-                return true;
-        }
-        return false;
-    }
-
-    public String getPrefix() {
-        return prefix;
-    }
-
-    public void setPrefix(String prefix) {
-        this.prefix = prefix;
-    }
-
-    private String[] cutArguments(String[] args, int startIndex, int endIndex) {
-        return Arrays.copyOfRange(args, startIndex, endIndex);
-    }
-
     protected abstract void handleImpl(MessageReceivedEvent event, String[] argStrings);
+
+    public List<String> getPermissions() {
+        return permission;
+    }
+
+    public int getMinArguments() {
+        return minArguments;
+    }
+
+    public void setMinArguments(int minArguments) {
+        this.minArguments = minArguments;
+    }
 }
