@@ -8,17 +8,13 @@ import kong.unirest.json.JSONArray;
 import kong.unirest.json.JSONObject;
 
 public class RecommendationHandler extends SpotifyApi {
-    private final String TRACKS_QUERY = "seed_tracks";
-    private final String GENRES_QUERY = "seed_genres";
-    private final String AVAILABLE_GENRE_SEEDS_RESPONSE_GENRES = "genres";
-    private final String RECOMMENDATIONS_RESPONSE_TRACKS = "tracks";
+    private static final String TRACKS_QUERY = "seed_tracks";
+    private static final String GENRES_QUERY = "seed_genres";
+    private static final String AVAILABLE_GENRE_SEEDS_RESPONSE_GENRES = "genres";
+    private static final String RECOMMENDATIONS_RESPONSE_TRACKS = "tracks";
     private final String recommendationsUrl = SpotifyUrlFactory.getRecommendationUrl();
     private final String availableGenreSeedsUrl = SpotifyUrlFactory.getAvailableGenreSeedsUrl();
     private TrackHandler trackHandler = new TrackHandler();
-
-    public RecommendationHandler() {
-
-    }
 
     public JSONArray getRecommendationsByTrack(String trackId) {
         return this.baseGetRequest(recommendationsUrl).queryString(TRACKS_QUERY, trackId).asJson().getBody().getArray();
@@ -26,16 +22,14 @@ public class RecommendationHandler extends SpotifyApi {
 
     public JSONArray getRecommendationsByTrackSearchQuery(String... searchQuery) {
         String trackId = trackHandler.getTrackIdBySearchQuery(searchQuery);
-        JSONArray recommendations = getRecommendationsByTrackIds(trackId);
-        return recommendations;
+        return getRecommendationsByTrackIds(trackId);
     }
 
     private JSONArray getRecommendationsByTrackIds(String... trackIds) {
         String seedsString = trimSeedArray(trackIds);
         JSONObject fetchResult = this.baseGetRequest(recommendationsUrl).queryString(TRACKS_QUERY, seedsString).asJson()
                 .getBody().getObject();
-        JSONArray recommendedTracks = fetchResult.getJSONArray("tracks");
-        return recommendedTracks;
+        return fetchResult.getJSONArray(RECOMMENDATIONS_RESPONSE_TRACKS);
     }
 
     private String trimSeedArray(String[] seedArray) {
@@ -58,7 +52,6 @@ public class RecommendationHandler extends SpotifyApi {
     public JSONArray getRecommendationsByGenre(String genre) {
         JSONObject response = this.baseGetRequest(recommendationsUrl).queryString(GENRES_QUERY, genre).asJson()
                 .getBody().getObject();
-        JSONArray recommendedTracks = response.getJSONArray(RECOMMENDATIONS_RESPONSE_TRACKS);
-        return recommendedTracks;
+        return response.getJSONArray(RECOMMENDATIONS_RESPONSE_TRACKS);
     }
 }
