@@ -1,6 +1,6 @@
 package audioplayer.commands;
 
-import main.Commands;
+import main.CommandManager;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import shared.commands.Command;
 
@@ -11,13 +11,12 @@ public class JumptoCommand extends Command {
         setDescription("sets the position of the current song");
         setTopic("music");
         setHelpString("<position in seconds to skip to in the current playing song>");
+        setMinArguments(1);
     }
 
     @Override
     protected void handleImpl(MessageReceivedEvent event, String[] argStrings) {
-        if (argStrings.length == 1) {
-            Commands.player.jumpto(event.getTextChannel(), Integer.parseInt(argStrings[1]));
-        }
+        CommandManager.player.jumpto(event.getTextChannel(), Integer.parseInt(argStrings[1]));
     }
 
 }
