@@ -1,6 +1,6 @@
 package audioplayer.commands;
 
-import main.Commands;
+import main.CommandManager;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import shared.commands.Command;
 
@@ -14,7 +14,7 @@ public class PauseCommand extends Command {
 
     @Override
     protected void handleImpl(MessageReceivedEvent event, String[] argStrings) {
-        Commands.player.togglePause(event.getTextChannel(), true);
+        CommandManager.player.togglePause(event.getTextChannel(), true);
     }
 
 }
