@@ -4,7 +4,7 @@ import java.util.Iterator;
 
 import com.sedmelluq.discord.lavaplayer.track.AudioTrack;
 
-import main.Commands;
+import main.CommandManager;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import shared.commands.Command;
 
@@ -19,12 +19,12 @@ public class QueueCommand extends Command {
     @Override
     protected void handleImpl(MessageReceivedEvent event, String[] argStrings) {
         StringBuilder queue = new StringBuilder("current queue: \n");
-        Iterator<AudioTrack> it = Commands.player.getQueue(event.getTextChannel()).iterator();
+        Iterator<AudioTrack> it = CommandManager.player.getQueue(event.getTextChannel()).iterator();
         int i = 1;
         while (it.hasNext()) {
             queue.append(i++).append(" ").append(it.next().getInfo().title).append("\n");
         }
-        Commands.sendBeautifulMessage(event, queue.toString());
+        CommandManager.sendBeautifulMessage(event, queue.toString());
     }
 
 }
