@@ -1,6 +1,6 @@
 package audioplayer.commands;
 
-import main.Commands;
+import main.CommandManager;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import shared.commands.Command;
 
@@ -11,20 +11,16 @@ public class RemoveCommand extends Command {
         setDescription("remove song from queue");
         setTopic("music");
         setHelpString("<number of song in queue>. use queue command to see queue");
+        setMinArguments(1);
     }
 
     @Override
     protected void handleImpl(MessageReceivedEvent event, String[] argStrings) {
-        if (argStrings.length < 1) {
-            main.Commands.sendMessage(event, getHelp());
-            return;
-        } else {
-            int indexToRemove = Integer.parseInt(argStrings[0]);
-            try {
-                Commands.player.remove(indexToRemove, event.getTextChannel());
-            } catch (NumberFormatException e) {
-                Commands.sendBeautifulMessage(event, "the position you have entered is invalid");
-            }
+        int indexToRemove = Integer.parseInt(argStrings[0]);
+        try {
+            CommandManager.player.remove(indexToRemove, event.getTextChannel());
+        } catch (NumberFormatException e) {
+            CommandManager.sendBeautifulMessage(event, "the position you have entered is invalid");
         }
     }
 
