@@ -1,6 +1,6 @@
 package audioplayer.commands;
 
-import main.Commands;
+import main.CommandManager;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import shared.commands.Command;
 
@@ -14,7 +14,7 @@ public class SkipCommand extends Command {
 
     @Override
     protected void handleImpl(MessageReceivedEvent event, String[] argStrings) {
-        Commands.player.skipTrack(event.getTextChannel());
+        CommandManager.player.skipTrack(event.getTextChannel());
     }
 
 }
