@@ -8,7 +8,7 @@ import org.simmetrics.metrics.StringDistances;
 import audioplayer.api.RecommendationHandler;
 import audioplayer.api.TrackHandler;
 import kong.unirest.json.JSONObject;
-import main.Commands;
+import main.CommandManager;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import shared.commands.Command;
 import shared.util.ListUtilies;
@@ -20,26 +20,22 @@ public class PlayRandomCommand extends Command {
         setDescription("plays random songs based on genre");
         setTopic("music");
         setHelpString("<genre>");
+        setMinArguments(1);
     }
 
     @Override
     protected void handleImpl(MessageReceivedEvent event, String[] argStrings) {
-        if (argStrings.length <= 0) {
-            Commands.sendMessage(event, getHelp());
-        } else {
-            String rawRequestedGenre = argStrings[0];
-            RecommendationHandler recommendationHandler = new RecommendationHandler();
-            List<String> availableGenres = recommendationHandler.getAvailableGenreSeeds();
-            List<JSONObject> recommendedTracks = null;
-            String requestedGenre = identifyRequestedGenre(rawRequestedGenre, availableGenres);
-            System.out.println(requestedGenre);
-            Commands.sendMessage(event, "Playing recommended tracks for genre: " + requestedGenre);
-            recommendedTracks = ListUtilies.castList(JSONObject.class,
-                    recommendationHandler.getRecommendationsByGenre(requestedGenre).toList());
-            if (recommendedTracks != null) {
-                recommendedTracks = recommendedTracks.subList(0, 2);
-                playRecommendedTracks(event, recommendedTracks);
-            }
+        String rawRequestedGenre = argStrings[0];
+        RecommendationHandler recommendationHandler = new RecommendationHandler();
+        List<String> availableGenres = recommendationHandler.getAvailableGenreSeeds();
+        List<JSONObject> recommendedTracks = null;
+        String requestedGenre = identifyRequestedGenre(rawRequestedGenre, availableGenres);
+        CommandManager.sendMessage(event, "Playing recommended tracks for genre: " + requestedGenre);
+        recommendedTracks = ListUtilies.castList(JSONObject.class,
+                recommendationHandler.getRecommendationsByGenre(requestedGenre).toList());
+        if (recommendedTracks != null) {
+            recommendedTracks = recommendedTracks.subList(0, 2);
+            playRecommendedTracks(event, recommendedTracks);
         }
     }
 
@@ -48,7 +44,7 @@ public class PlayRandomCommand extends Command {
             String firstArtistName = TrackHandler
                     .getNameOfArtist(TrackHandler.getArtistsOfTrack(recommendedTrack).getJSONObject(0));
             String trackName = TrackHandler.getNameOfTrack(recommendedTrack);
-            Commands.player.loadAndPlay(event, "ytsearch: " + firstArtistName + " " + trackName);
+            CommandManager.player.loadAndPlay(event, "ytsearch: " + firstArtistName + " " + trackName);
         });
     }
 
