@@ -1,6 +1,6 @@
 package audioplayer.commands.playlist;
 
-import static main.Commands.sendBeautifulMessage;
+import static main.CommandManager.sendBeautifulMessage;
 
 import com.sedmelluq.discord.lavaplayer.player.AudioLoadResultHandler;
 import com.sedmelluq.discord.lavaplayer.tools.FriendlyException;
@@ -11,7 +11,7 @@ import database.PlaylistTable;
 import database.Song;
 import database.SongPlaylistTable;
 import database.SongTable;
-import main.Commands;
+import main.CommandManager;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import shared.commands.Command;
 
@@ -22,57 +22,55 @@ public class AddToPlaylistCommand extends Command {
         setDescription("Searches for a song with the keywords and adds it to the given playlist");
         setTopic("music");
         setHelpString("<playlist name> <keywords>");
+        setMinArguments(2);
     }
 
     @Override
     protected void handleImpl(MessageReceivedEvent event, String[] argStrings) {
-        if (argStrings.length >= 2) {
-            String playlistNameToAddTo = argStrings[0];
-            String search = "ytsearch: ";
-            for (int i = 1; i < argStrings.length; i++) {
-                search += argStrings[i] + " ";
-            }
-            Commands.player.fetchAudioTrack(event.getTextChannel(), search, new AudioLoadResultHandler() {
-                @Override
-                public void trackLoaded(AudioTrack track) {
-                    if (!SongTable.hasSong(track.getInfo().uri))
-                        SongTable.insertSong(new Song(track.getInfo().title, track.getInfo().uri));
-                    Song song = SongTable.getSongsByUrl(track.getInfo().uri).get(0);
-                    if (SongPlaylistTable.insertSongIntoPlaylist(song,
-                            SongPlaylistTable.getPlaylistByName(playlistNameToAddTo)) != 0) {
-                        sendBeautifulMessage(event,
-                                "added \"" + track.getInfo().title + "\" to playlist " + playlistNameToAddTo);
-                    } else {
-                        sendBeautifulMessage(event, "there is no playlist " + playlistNameToAddTo);
-                    }
-                }
-
-                @Override
-                public void playlistLoaded(AudioPlaylist playlist) {
-                    if (!SongTable.hasSong(playlist.getTracks().get(0).getInfo().uri))
-                        SongTable.insertSong(new Song(playlist.getTracks().get(0).getInfo().title,
-                                playlist.getTracks().get(0).getInfo().uri));
-                    Song song = SongTable.getSongsByUrl(playlist.getTracks().get(0).getInfo().uri).get(0);
-                    if (SongPlaylistTable.insertSongIntoPlaylist(song,
-                            PlaylistTable.getPlaylist(playlistNameToAddTo)) != 0) {
-                        sendBeautifulMessage(event,
-                                "added \"" + song.getTitle() + "\" to playlist " + playlistNameToAddTo);
-                    } else {
-                        sendBeautifulMessage(event, "there is no playlist " + playlistNameToAddTo);
-                    }
-                }
-
-                @Override
-                public void noMatches() {
-                    sendBeautifulMessage(event, "Nothing found for keywords");
-                }
-
-                @Override
-                public void loadFailed(FriendlyException exception) {
-                    System.out.println("loadFailed");
-                }
-            });
+        String playlistNameToAddTo = argStrings[0];
+        String search = "ytsearch: ";
+        for (int i = 1; i < argStrings.length; i++) {
+            search += argStrings[i] + " ";
         }
+        CommandManager.player.fetchAudioTrack(event.getTextChannel(), search, new AudioLoadResultHandler() {
+            @Override
+            public void trackLoaded(AudioTrack track) {
+                if (!SongTable.hasSong(track.getInfo().uri))
+                    SongTable.insertSong(new Song(track.getInfo().title, track.getInfo().uri));
+                Song song = SongTable.getSongsByUrl(track.getInfo().uri).get(0);
+                if (SongPlaylistTable.insertSongIntoPlaylist(song,
+                        SongPlaylistTable.getPlaylistByName(playlistNameToAddTo)) != 0) {
+                    sendBeautifulMessage(event,
+                            "added \"" + track.getInfo().title + "\" to playlist " + playlistNameToAddTo);
+                } else {
+                    sendBeautifulMessage(event, "there is no playlist " + playlistNameToAddTo);
+                }
+            }
+
+            @Override
+            public void playlistLoaded(AudioPlaylist playlist) {
+                if (!SongTable.hasSong(playlist.getTracks().get(0).getInfo().uri))
+                    SongTable.insertSong(new Song(playlist.getTracks().get(0).getInfo().title,
+                            playlist.getTracks().get(0).getInfo().uri));
+                Song song = SongTable.getSongsByUrl(playlist.getTracks().get(0).getInfo().uri).get(0);
+                if (SongPlaylistTable.insertSongIntoPlaylist(song,
+                        PlaylistTable.getPlaylist(playlistNameToAddTo)) != 0) {
+                    sendBeautifulMessage(event, "added \"" + song.getTitle() + "\" to playlist " + playlistNameToAddTo);
+                } else {
+                    sendBeautifulMessage(event, "there is no playlist " + playlistNameToAddTo);
+                }
+            }
+
+            @Override
+            public void noMatches() {
+                sendBeautifulMessage(event, "Nothing found for keywords");
+            }
+
+            @Override
+            public void loadFailed(FriendlyException exception) {
+                System.out.println("loadFailed");
+            }
+        });
     }
 
 }
