@@ -1,6 +1,6 @@
 package audioplayer.commands.playlist;
 
-import static main.Commands.sendBeautifulMessage;
+import static main.CommandManager.sendBeautifulMessage;
 
 import database.Playlist;
 import database.PlaylistTable;
@@ -16,24 +16,21 @@ public class SaveHistoryCommand extends Command {
         setDescription("Creates a new playlist with the given name and adds the last 10 songs from history to it");
         setTopic("music");
         setHelpString("<playlist name>");
+        setMinArguments(1);
     }
 
     @Override
     protected void handleImpl(MessageReceivedEvent event, String[] argStrings) {
-        if (argStrings.length == 1) {
-            String playlistName = argStrings[0];
-            if (PlaylistTable.getPlaylist(playlistName) != null) {
-                sendBeautifulMessage(event, "playlistname already exists");
-                return;
-            } else {
-                PlaylistTable.createPlaylist(playlistName);
-                final Playlist playlist = PlaylistTable.getPlaylist(playlistName);
-                SongHistoryTable.getLastSongs().forEach((id, song) -> {
-                    SongPlaylistTable.insertSongIntoPlaylist(song, playlist);
-                });
-                sendBeautifulMessage(event, "saved playlist " + playlistName);
-
-            }
+        String playlistName = argStrings[0];
+        if (PlaylistTable.getPlaylist(playlistName) != null) {
+            sendBeautifulMessage(event, "playlistname already exists");
+        } else {
+            PlaylistTable.createPlaylist(playlistName);
+            final Playlist playlist = PlaylistTable.getPlaylist(playlistName);
+            SongHistoryTable.getLastSongs().forEach((id, song) -> {
+                SongPlaylistTable.insertSongIntoPlaylist(song, playlist);
+            });
+            sendBeautifulMessage(event, "saved playlist " + playlistName);
         }
     }
 
