@@ -1,7 +1,7 @@
 package imageboards.commands;
 
 import imageboards.api.PixabayHandler;
-import main.Commands;
+import main.CommandManager;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import shared.commands.Command;
 
@@ -13,20 +13,16 @@ public class PixabayCommand extends Command {
         setDescription("search for any pictures you want");
         setTopic("images");
         setHelpString("<search term to search for>");
+        setMinArguments(1);
     }
 
     @Override
     protected void handleImpl(MessageReceivedEvent event, String[] argStrings) {
-        if (argStrings.length >= 1) {
-            String searchQuery = "";
-            for (String subString : argStrings) {
-                searchQuery += subString + " ";
-            }
-            PixabayHandler pixabayHandler = new PixabayHandler();
-            String imageUrl = pixabayHandler.getRandomImageUrlByQuery(searchQuery);
+        String searchQuery = String.join(" ", argStrings);
+        PixabayHandler pixabayHandler = new PixabayHandler();
+        String imageUrl = pixabayHandler.getRandomImageUrlByQuery(searchQuery);
 
-            Commands.sendMessage(event, imageUrl);
-        }
+        CommandManager.sendMessage(event, imageUrl);
     }
 
 }
