@@ -1,7 +1,7 @@
 package imageboards.commands;
 
 import imageboards.api.NekobotHandler;
-import main.Commands;
+import main.CommandManager;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import shared.commands.Command;
 
@@ -17,7 +17,7 @@ public class ThighCommand extends Command {
     protected void handleImpl(MessageReceivedEvent event, String[] argStrings) {
         NekobotHandler nekobotHandler = new NekobotHandler();
         String imageUrl = nekobotHandler.getThighImage();
-        Commands.sendMessage(event, imageUrl);
+        CommandManager.sendMessage(event, imageUrl);
     }
 
 }
